Feature: Quit command
The Context class is used to encapsulate the common data and state of
the application. Since the context's lifetime is guaranteed to be for
the entire duration of the program, it is cached in the Command upon
instantiation.

This is commonly referred to as the "Context Pattern". It provides a
way to pass around a shared "context" and avoids using a singleton.